feat: 在线订单基本统计-修改

+++ catguild-business/src/main/java/cn/catguild/business/erp/application/dto/StatisticsChart.java
@@ -41,7 +41,7 @@ public class StatisticsChart {
     /**
      * 图表侧边：销售量
      */
-    private Long salesVolume;
+    private BigDecimal salesVolume;
 
 
 }
